WIP: LittleFS boot loop workaround - disable auto-format on mount failure
- Modified initFileSystem() to not attempt auto-format (causes error 261 flash op fail)
- Added CONFIG_LITTLEFS_BLOCK_SIZE=4096 to platformio.ini to match ESP32-S3 flash sector size
- Device now boots to AP mode without LittleFS mount (web UI unavailable but no boot loop)
- Next: Convert HTML/CSS to embedded PROGMEM to eliminate filesystem dependency

Issues addressed:
- LittleFS partition corruption with error 261 (ESP_ERR_FLASH_OP_FAIL)
- NVS initialization failures with error 4363 (ESP_ERR_NVS_NOT_INITIALIZED)
- Infinite boot loop on WiFi initialization failure

+++ src/logger.cpp
@@ -128,10 +128,10 @@ void logEvent(LogLevel level, const String &message) {
     Serial.print(emoji);
     Serial.print(" ");
     Serial.print(timeStr);
-    Serial.printf(" [ %lums] %s\n", ms, messageText.c_str());
+    Serial.printf(" [ %lums] %s\n", static_cast<unsigned long>(ms), messageText.c_str());
   } else {
     Serial.print(timeStr);
-    Serial.printf(" [ %lums] %s\n", ms, message.c_str());
+    Serial.printf(" [ %lums] %s\n", static_cast<unsigned long>(ms), message.c_str());
   }
 
   // Store in ring buffer for web UI.
